% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ColourText.R
\name{ColourText}
\alias{ColourText}
\title{ColourText}
\usage{
ColourText(x, color)
}
\arguments{
\item{x}{Text to be coloured in either HTML or LaTeX}

\item{color}{Colour to be used}
}
\value{
Returns the coloured text to be used in-line in RMD documents
}
\description{
ColourText
}
\examples{
To be used in-line in Rmd Documents
# `r ColourText("MY RED TITLE",'red')`
`r ColourText("MY RED TEXT",'red')`


}
