% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpi_adjust.R
\name{cpi_adjust}
\alias{cpi_adjust}
\title{Adjust by cpi, i.e. set to current prices}
\usage{
cpi_adjust(df, vars, timevar, at_date = NULL)
}
\arguments{
\item{vars}{Variables to adjust}

\item{timevar}{Date variable}

\item{at_date}{Date to adjust to}
}
\value{
df with adjusted series
}
\description{
Adjust by cpi, i.e. set to current prices
}
\examples{
yfirverd <- yfirverd_get_data()

yfirverd \%>\%
  sql_clean() \%>\%
  cpi_adjust(
  vars = vars(Kaupverd, AuglystSoluverd),
  timevar = Dim_Timi_Utgefid
  )

}
