% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenerateDocumentation.R
\name{docstring_to_roxygen}
\alias{docstring_to_roxygen}
\title{docstring_to_roxygen}
\usage{
docstring_to_roxygen(
  fun,
  fun_name = as.character(substitute(fun)),
  default_title = "Title not detected",
  error = TRUE
)
}
\arguments{
\item{fun}{a function to document ( will error if no r oxygen comments in it)}

\item{fun_name}{defaults to a string of the function name, otherwise defines the name the documentaion file will be named}

\item{default_title}{If you want to change the title text (untested)}
}
\value{
The function returns the text from r oxygen comments
}
\description{
This function has been stolen from the docsting package.
It takes a function and returns the text of the ROxygen comments erros if there are none.
}
