% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenerateDocumentation.R
\name{GenerateDocumentation}
\alias{GenerateDocumentation}
\title{GenerateDocumentation}
\usage{
GenerateDocumentation(
  fun,
  fun_name = as.character(substitute(fun)),
  rstudio_pane = getOption("docstring_rstudio_help_pane"),
  default_title = "Title not detected",
  DocDir = paste0(here::here(), "/HMS_Economics/Reference/AnalysisDocumentation/")
)
}
\arguments{
\item{fun}{a function to document ( will error if no r oxygen comments in it)}

\item{fun_name}{defaults to a string of the function name, otherwise defines the name the documentaion file will be named}

\item{rstudio_pane}{to write the help to rstudio}

\item{default_title}{If you want to change the title text (untested)}

\item{DocDir}{a directory for the documentaion to reside in (defaults to the documentation dir within the HMS_DS Git repo, otherwise this will need to be set for this to work)}
}
\value{
The function returns nothing but should write pdf documentation to the directory specified
}
\description{
This function with auto generate pdf documentaion. It requires Roxygen comments to be **inside** the function to be documented
It also requires Miktex to be installed correctly.
It will also allow documented functions to appear in R help using help(function_name) or ?function_name
It currently writes the Documentation to the HMS_DS directory structure in HMS_DS/HMS_Economics/Reference/AnalysisDocumentation/
However it will take an argument DocDir to alter this.
Code has been stolen and edited from the docsting package.
}
\note{
It will currently understand the following tags:

return description to describe the output that a function creates.

title, description, details are available if you want to be more explicit in the sections. This also allows you to have a description section longer than one paragraph if you explicitly use the description keyword.

param list the parameters used in the function

note Creates a note section

section SectionTitle: You can create your own sections using the section keyword. Following section must be the section title in sentence case and must be finished with a colon. Subsections do not have their own keyword but can be added using the Rd backslash subsection{} command.

usage If you want to have more control over the usage section you can either overwrite the generated usage or add to it by using this keyword

references If you want to add references to your documentation. This is important (and appreciated) especially when you are writing a function that implements an algorithm in a journal article. Having a reference to the source is vital in those situations.

examples which will create an examples section in the documentation. Using docstring there isn't a nice way to run the examples like you can using the example() function but they will still show up in the documentation which is sometimes the most important part. Users can still copy/paste if they desire.

Comments do not need to be consecutive they could be throughout the function.
}
\examples{
# define a function with R oxygen comments
#' square <- function(x){
#' Square a number

#' Calculates the square of the input

#' param x the input to be squared
 return(x^2)
}

Call the GenerateDocumentation function
GenerateDocumentation(square)

}
