% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_connectons.R
\name{get_connection_JDBC}
\alias{get_connection_JDBC}
\title{get_connection_JDBC Connects to a database server using JDBC}
\usage{
get_connection_JDBC(
  dbhost = "sql02",
  dbname = "HgrDwh",
  dbcp = "D:/R/Drivers/mssql-jdbc-7.2.1.jre8.jar",
  user,
  password
)
}
\arguments{
\item{dbhost}{Hostname name to connect to}

\item{dbname}{Databse name to connect to}

\item{dbcp}{Class path for driver}

\item{user}{Username}

\item{password}{Password}
}
\value{
DBI conection object
}
\description{
get_connection_JDBC Connects to a database server using JDBC
}
