% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_connectons.R
\name{get_connection_windows}
\alias{get_connection_windows}
\title{get_connection_windows}
\usage{
get_connection_windows(db = c("HgrDwh_Prod", "HgrDev"))
}
\value{

}
\description{
get_connection_windows
}
