% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_connectons.R
\name{get_tbl}
\alias{get_tbl}
\title{Access specific table from the datawarehouse.}
\usage{
get_tbl(tbl, con = get_connection_windows(), schema = "dwh")
}
\arguments{
\item{tbl}{Name of the table in the datawarehouse}

\item{con}{Connection to the datawarehouse}

\item{schema}{}
}
\value{
tbl object to use with warehouse. See https://dbplyr.tidyverse.org/
}
\description{
Thin wrapper around tbl from dbplyr. Allows the user to use some of the common
dplyr function on an object without importing it to R.
}
\examples{
visitolur <- get_tbl(Fact_Visitolur)

visitolur \%>\%
  filter(dim_visitolur == 3) \%>\%
  collect()
}
