% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenerateDocumentation.R
\name{has_docstring}
\alias{has_docstring}
\title{has_docstring}
\usage{
has_docstring(fun, fun_name = as.character(substitute(fun)))
}
\arguments{
\item{fun}{a function to document ( will error if no r oxygen comments in it)}

\item{fun_name}{defaults to a string of the function name, otherwise defines the name the documentaion file will be named}
}
\value{
The function returns the text from r oxygen comments
}
\description{
This function has been stolen from the docsting package.
It checks the function for roxygen type comments (any no order necessary)
}
