% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index.R
\name{index}
\alias{index}
\title{Indexing a time series}
\usage{
index(val, time_var, at_time = min(time_var[!is.na(val)], na.rm = TRUE))
}
\arguments{
\item{val}{a numeric variable to scale}

\item{time_var}{time variable or other indexing variable}

\item{at_time}{the time or index value at the base year}
}
\value{
A scaled series.
}
\description{
Scales a time series in such a way that it gets the value 100 at a given point in time
}
\examples{
economics_long \%>\%
 group_by(variable) \%>\%
 mutate(val = index(value, date))


}
