% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seasonal_adjustment.R
\name{nested_seas}
\alias{nested_seas}
\title{Seasonal Adjustment (map)}
\usage{
nested_seas(data, var, date, frequency = 12, ...)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Seasonal Adjustment (map)
}
