% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_hagstofan.R
\name{read_hagstofan}
\alias{read_hagstofan}
\title{Read data from Hagstofa Íslands (Statistics Iceland)}
\usage{
read_hagstofan(
  file,
  na = "..",
  encoding = "WINDOWS-1252",
  decimal_mark = ",",
  grouping_mark = ".",
  delim = ";",
  ...
)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Read data from Hagstofa Íslands (Statistics Iceland)
}
\examples{

}
