% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_hagstofan.R
\name{read_hagstofan_alt}
\alias{read_hagstofan_alt}
\title{Read data from Hagstofa Íslands (Statistics Iceland)}
\usage{
read_hagstofan_alt(
  file,
  na = "..",
  encoding = "WINDOWS-1252",
  decimal_mark = ".",
  grouping_mark = ",",
  delim = ";",
  ...
)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
The same as read_hagstofan, but with dots as decimal marks.
}
