% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/the_answer_to_live_universe_and_everything.R
\name{the_answer_to_life_universe_and_everything}
\alias{the_answer_to_life_universe_and_everything}
\title{The Answare to Live, the Universe and Everything}
\usage{
the_answer_to_life_universe_and_everything()
}
\value{

}
\description{
The Answare to Live, the Universe and Everything
}
