% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_hms.R
\name{theme_flip}
\alias{theme_flip}
\title{Use vertical gridlines instead of horizontal}
\usage{
theme_flip(color = gray, ...)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Use vertical gridlines instead of horizontal
}
