% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_hms.R
\name{theme_hms}
\alias{theme_hms}
\title{Theme for [HMS](https://www.hms.is/)}
\usage{
theme_hms(fsm = 1)
}
\arguments{
\item{fsm}{Font Size Multiplier}
}
\value{
theme
}
\description{
Theme for [HMS](https://www.hms.is/)
}
\examples{

cars \%>\%
ggplot(aes(speed, dist), fill = blue) +
 geom_col() +
 labs(
   title = "Some title",
   subtitle = "Some subtitle",
   caption = "Some caption"
 ) +
 theme_hms()
}
