% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_hms.R
\name{theme_set_hms}
\alias{theme_set_hms}
\title{Set the HMS theme as the default theme}
\usage{
theme_set_hms(theme = theme_hms(), overwrite_default_palette = TRUE)
}
\arguments{
\item{overwrite_default_palette}{}
}
\value{

}
\description{
Set the HMS theme as the default theme
}
