% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_hms.R
\docType{data}
\name{width_wide_report}
\alias{width_wide_report}
\title{Title}
\format{An object of class \code{numeric} of length 1.}
\usage{
width_wide_report
}
\value{

}
\description{
Title
}
\keyword{datasets}
