% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yfirverd.R
\name{yfirverd}
\alias{yfirverd}
\title{Title}
\usage{
yfirverd(
  ...,
  connection = get_connection_windows(),
  print_plot = TRUE,
  eftir = NULL,
  filters = NULL
)
}
\arguments{
\item{...}{Grouping variables (unquoted), used for faceting}

\item{connection}{A connection to the HMS database}

\item{eftir}{Name of the groups to include in the title}
}
\value{
The data data used for the plot.
}
\description{
Title
}
