% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_connectons.R
\name{get_connection_windows}
\alias{get_connection_windows}
\title{get_connection_windows}
\usage{
get_connection_windows(db = "HgrDwh_Prod")
}
\arguments{
\item{db}{Name of the dsn, either "HgrDwh_Prod" or "HgrDev"}
}
\value{
a connection to the HMS database
}
\description{
get_connection_windows
}
